js: change names, add reload limit
