--- js/popcorn/src/config.ts
@@ -1,4 +1,0 @@
-export const INIT_VM_TIMEOUT_MS = 30_000;
-export const CALL_TIMEOUT_MS = 60_000;
-export const HEARTBEAT_TIMEOUT_MS = 60_000;
-export const HEARTBEAT_INTERVAL_MS = 500;
