+++ js/popcorn/src/types.ts
@@ -1,3 +1,9 @@
+export const INIT_VM_TIMEOUT_MS = 30_000;
+export const CALL_TIMEOUT_MS = 60_000;
+export const HEARTBEAT_TIMEOUT_MS = 60_000;
+export const HEARTBEAT_INTERVAL_MS = 500;
+export const MAX_RELOAD_N = 3;
+
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export type AnySerializable = any;
